Scripts merged
shodan_bulk_search.py and shodanparse.py merged in one script for more automatized process. Also, execution from terminal now available.
