+++ shodan_bulk_search.py
@@ -1,12 +1,26 @@
+import csv
 import shodan
 import time
 import pickle
+import argparse
 from tqdm import tqdm
+
+
+def take_second(elem):
+    return elem[1]
+
+
+"""Argparse for terminal execution"""
+ap = argparse.ArgumentParser()
+ap.add_argument("-l", "--list", required=True, help="Txt file that has list of IP addresses.")
+args = vars(ap.parse_args())
+"""END argparse for terminal execution"""
 
 api = shodan.Shodan('APIKEY')
 
 ip_list = []
 processed_ips = set()
+output = ""
 
 try:
     with open("processed_ips", "rb") as fp:  # for handling many IPs, this way processed IPs will be stored and reloaded
@@ -18,7 +32,6 @@ with open("ip_list", "r", encoding="utf-8") as fp:  # IP list should be provided
     for line in fp:
         ip_list.append(line.split(" ")[1].strip())
 
-with open("output_file", "a", encoding="UTF-8") as fp:  # output file
     for ip in tqdm(ip_list):
         if ip in processed_ips:
             continue
@@ -37,7 +50,7 @@ with open("output_file", "a", encoding="UTF-8") as fp:  # output file
                 asn = info["asn"]
             except KeyError as e:
                 asn = "none"
-            fp.write("{},{},{},{},{}\n".format(info["ip_str"],
+            output += ("{} {} {} {} {}\n".format(info["ip_str"],
                                                info["os"],
                                                asn,
                                                " ".join(str(port) for port in info["ports"]),
@@ -51,3 +64,31 @@ with open("output_file", "a", encoding="UTF-8") as fp:  # output file
                 processed_ips.add(ip)
                 pickle.dump(processed_ips, pfp)
             print("For IP:" + ip + " this error is produced:" + str(e))
+
+    first_line = output.partition('\n')[0]
+    column_number = first_line.count(' ') + 1
+    for idx in range(column_number):
+        words = []
+        csv_words = []
+
+        for row in output.splitlines():
+            row = row.split(" ")
+            csv_words = row[idx].split(" ")
+
+            for i in csv_words:
+                words.append(i)
+
+        words_counted = []
+        for i in words:
+            x = words.count(i)
+            words_counted.append((i, x))
+        count_set = set(words_counted)
+
+        """Sorted by descending order of 2nd element which is occurence count and writes to csv"""
+        with open(str(idx) + ".csv", 'w', newline='') as f:
+            writer = csv.writer(f)
+            writer.writerows(sorted(count_set, key=take_second,
+                                    reverse=True))
+
+
+
